fix: remove narwhal intake outtake from robot container that was preventing code from building. move intake outtake state to utils to be consistant.
