+++ src/main/java/frc/robot/util/NarwhalIntakeOuttakeState.java
@@ -1,4 +1,4 @@
-package frc.robot.subsystems.narwhal;
+package frc.robot.util;
 
 /**
  * <h2> NarwhalIntakeOuttakeState </h2>
